op-e2e: Use helper to create dishonest helper.
Reduces the amount of boilerplate required in the actual test.
